correct types in strcspn strspn

+++ include/ft_string.h
@@ -6,7 +6,7 @@
 /*   By: ecarvalh <ecarvalh@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/23 00:29:52 by ecarvalh          #+#    #+#             */
-/*   Updated: 2024/02/12 19:10:02 by ecarvalh         ###   ########.fr       */
+/*   Updated: 2024/02/26 16:16:23 by ecarvalh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,8 +17,8 @@ char	*ft_strdup(const char *s);
 char	*ft_strchr(const char *s, int c);
 char	*ft_strrchr(const char *s, int c);
 char	*ft_strnstr(const char *s1, const char *s2, size_t n);
-int		ft_strspn(char const *str, char const *accept);
-int		ft_strcspn(char const *str, char const *reject);
+size_t	ft_strspn(char const *str, char const *accept);
+size_t	ft_strcspn(char const *str, char const *reject);
 char	*ft_strtok(char *str, char const *delim);
 size_t	ft_strlen(const char *s);
 size_t	ft_strlcpy(char *dst, const char *src, size_t sz);
